--wip-- [skip ci]

+++ src/main/scala/com/github/celadari/jsonlogicscala/evaluate/defaults/OperatorRightShiftZeroFillerBitwise.scala
@@ -7,17 +7,17 @@ object OperatorRightShiftZeroFillerBitwise extends Operator {
   def $greater$greater$greater(num1: java.lang.Byte, num2: java.lang.Byte): java.lang.Integer = num1 >>> num2.toInt
   def $greater$greater$greater(num1: java.lang.Byte, num2: java.lang.Short): java.lang.Integer = num1 >>> num2.toInt
   def $greater$greater$greater(num1: java.lang.Byte, num2: java.lang.Integer): java.lang.Integer = num1 >>> num2
-  def $greater$greater$greater(num1: java.lang.Byte, num2: java.lang.Long): java.lang.Integer = num1 >>> num2
+  def $greater$greater$greater(num1: java.lang.Byte, num2: java.lang.Long): java.lang.Integer = num1 >>> num2.toInt
 
   def $greater$greater$greater(num1: java.lang.Short, num2: java.lang.Byte): java.lang.Integer = num1 >>> num2.toInt
   def $greater$greater$greater(num1: java.lang.Short, num2: java.lang.Short): java.lang.Integer = num1 >>> num2.toInt
   def $greater$greater$greater(num1: java.lang.Short, num2: java.lang.Integer): java.lang.Integer = num1 >>> num2
-  def $greater$greater$greater(num1: java.lang.Short, num2: java.lang.Long): java.lang.Integer = num1 >>> num2
+  def $greater$greater$greater(num1: java.lang.Short, num2: java.lang.Long): java.lang.Integer = num1 >>> num2.toInt
 
   def $greater$greater$greater(num1: java.lang.Integer, num2: java.lang.Byte): java.lang.Integer = num1 >>> num2.toInt
   def $greater$greater$greater(num1: java.lang.Integer, num2: java.lang.Short): java.lang.Integer = num1 >>> num2.toInt
   def $greater$greater$greater(num1: java.lang.Integer, num2: java.lang.Integer): java.lang.Integer = num1 >>> num2
-  def $greater$greater$greater(num1: java.lang.Integer, num2: java.lang.Long): java.lang.Integer = num1 >>> num2
+  def $greater$greater$greater(num1: java.lang.Integer, num2: java.lang.Long): java.lang.Integer = num1 >>> num2.toInt
 
   def $greater$greater$greater(num1: java.lang.Long, num2: java.lang.Byte): java.lang.Long = num1 >>> num2.toInt
   def $greater$greater$greater(num1: java.lang.Long, num2: java.lang.Short): java.lang.Long = num1 >>> num2.toInt
